laba2

+++ app/Http/Controllers/HouseController.php
@@ -2,7 +2,9 @@
 
 namespace App\Http\Controllers;
 use App\Models\House;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Gate;
 
 
@@ -14,10 +16,13 @@ class HouseController extends Controller
         return view('house.index',data: compact('house'));
     }
     public function create() {
+
         return view('house.create');
     }
     public function edit(House $house) {
-
+        if(! Gate::allows('house_update', $house)) {
+            return response('You cannot delete, not your profile', 483);
+        }
         return view('house.edit',compact('house'));
     }
     public function store() {
@@ -29,11 +34,15 @@ class HouseController extends Controller
             'quantity_of_characters' => 'required',
             'quantity_of_live_characters'=> 'required',
         ]);
+        $data['user_id']=Auth::id();
         House::create($data);
 
         return redirect('/');
     }
     public function update(House $house) {
+        if(! Gate::allows('house_update', $house)) {
+            return response('You cannot delete, not your profile', 483);
+        }
         $valid = request()->validate([
             'emblem' => 'required',
             'name' => 'required',
@@ -47,9 +56,19 @@ class HouseController extends Controller
 
     }
     public function destroy(House $house) {
-
+        if(! Gate::allows('house_update', $house)) {
+            return response('You cannot delete, not your profile', 483);
+        }
         $house->delete();
         return redirect()->route('houses');
+    }
+    public function forceDelete($house)
+    {
+        $houseRelease = \App\Models\House::find($house);
+        if (!Gate::allows('force_delete', $houseRelease)) abort(403);
+        $user = User::find($houseRelease->user_id);
+        $houseRelease->forceDelete();
+        return redirect('/');
     }
     public function show(House $house) {
         return view('house.show', compact('house'));
